Changed int to double.

+++ ParsingValuesFromString/src/Main.java
@@ -4,7 +4,7 @@ public class Main {
         String numberAsString = "2020";
         System.out.println("numberAsString = " + numberAsString);
 
-        int number = Integer.parseInt(numberAsString);
+        double number = Double.parseDouble(numberAsString);
         System.out.println("Number = " + number);
 
         numberAsString += 1;
